+++ ext/stub/instance.zep.c
@@ -82,7 +82,7 @@ PHP_METHOD(Stub_Instance, testIssue1339)
 	zephir_array_fast_append(&parameters, &_0);
 	ZEPHIR_INIT_NVAR(&_0);
 	object_init_ex(&_0, stub_arrayobject_ce);
-	ZEPHIR_CALL_METHOD(NULL, &_0, "__construct", NULL, 56);
+	ZEPHIR_CALL_METHOD(NULL, &_0, "__construct", NULL, 57);
 	zephir_check_call_status();
 	zephir_array_fast_append(&parameters, &_0);
 	ZEPHIR_INIT_NVAR(&_0);